Translate most of the assembly over to C (and extras)
This is a big commit because I was too lazy to split it into multiple
because of how many changes there are.

This obviously translates most of the assembly over to C. Keep in mind
that most of the translation is pretty direct, so expect to see a lot of
jank in the code.

This also reverts some Hawkeye specific changes, mostly to maintain
purity with the original source.

This also fixes a bug in the memory allocation scheme they use for
resources which screws up on any architecture that's not 32-bit,
therefore allowing you to run this on x86_64.

Some of the assembly code currently remains untranslated, which I will
get around it in the next few commits.

Signed-off-by: bjrkk <cfxqpkjq@gmail.com>